Avoid using bsg_tiles_{X,Y} in host code
I've now defined TILES_X and TILES_Y constants within each example
instead of providing these with `-D` flags in the Makefile. I know this
is more repetitive, but it emphasizes that the host code doesn't
actually need these macros---the size of the tile group that the host
code uses is a run-time decision, and it doesn't need to be baked in.

I also wanted to emphasize that we are not required to use this precise
constant name, unlike in the device code.